Lotta stuff added (Mainly sprite for rect)
Right now the crate's on circular collision, but a basic representation of the rectangle following the circle has been made. Classes like DebugSprite aren't used right now, but possibly later (will probably remove when this is ready for release). Next up is the collision, which will take a good bit probably. Also, sorry about code if you're reading this, it's a bit of a mess right now lmao.