"Fixed" Collision Detection/Basic Collision
"Fixed" the collision detection, for some reason the check was checking 3 tiles above the actual tiles and 2 tiles to the right of the actual tiles? Not sure why, but I just added them back so I'm hoping it's a good fix, but if I'm being honest it might come up as a problem later. Also added basic collision, but having problems with it sinking into the tile before applying collision code which makes it slowly rise out of the tile and also I think negates the reflection velocity being applied at first. Need to figure out how to fix that /: